0.7.2: Handle primary property rendering errors
This takes care of broken property rendering in the primary rendering function. Broken properties are now rendered as strings indicating the error with the property, and the resource-state is changed to ‘corrupt’.
Note that this DOES NOT handle issues with any secondary rendering that may take place. Certain resources have external rendering functions for various reasons (i.e. to decorate with data from an external system). We’ll have to go resource by resource to deal with that.