Modify root initialization on startup
Meta will no longer ‘fail’ if it can’t determine who the root user is on startup.  There is an issue that prevents meta from talking to security until /bootstrap is called.  Failing on startup prevents /bootstrap from being called.  The solution is no not fail, to simply log the fact that there is no root user when that is the case.  Note that before the call to bootstrap, there’s nothing you can do with meta, so it doesn’t matter, and once bootstrap *is* called, a root user will be set.
Added a boolean `root_initialized` field to output of `/about` endpoint.