Create kube namespace if needed on chart deploy
This handles the situation where the Meta environment does not yet have an associated kubernetes namespace (named from environment ID). The UI controls the namespace that’s used so there’s no concern about ad-hoc namespace creation from there. In future we can add a query param to control whether or not a namespace should be created or not.