Fix issue with deleting ‘failed’ appdeployment
This is a crude patch to move us forward. The code really needs to look into “why” the deployment failed so it can clean up appropriately. Right now, if it’s seen that the deployment failed for any reason, everything is ignored and the appdeployment is forcibly removed from Meta. This could leave orphans both in and out of meta.

Also added a validation to force ‘release-name’ to be dns-1123 compliant.